Add OptimizerRuleScatterInCluster.h and cpp files; Remove the rules from OptimizerRules.cpp
