+++ arangod/Aql/OptimizerRules.h
@@ -84,11 +84,6 @@ void clusterPushSubqueryToDBServer(Optimizer* opt,
                                    OptimizerRule const& rule);
 #endif
 
-/// @brief scatter operations in cluster - send all incoming rows to all remote
-/// clients
-void scatterInClusterRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
-                          OptimizerRule const&);
-
 #ifdef USE_ENTERPRISE
 void distributeOffsetInfoToClusterRule(aql::Optimizer* opt,
                                        std::unique_ptr<aql::ExecutionPlan> plan,
@@ -132,26 +127,6 @@ void replaceLikeWithRangeRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
 void replaceEntriesWithObjectIteration(Optimizer*,
                                        std::unique_ptr<ExecutionPlan>,
                                        OptimizerRule const&);
-
-void createScatterGatherSnippet(
-    ExecutionPlan& plan, TRI_vocbase_t* vocbase, ExecutionNode* node,
-    bool isRootNode, std::vector<ExecutionNode*> const& nodeDependencies,
-    std::vector<ExecutionNode*> const& nodeParents,
-    SortElementVector const& elements, size_t numberOfShards,
-    std::unordered_map<ExecutionNode*, ExecutionNode*> const& subqueries,
-    Collection const* collection);
-
-//// @brief enclose a node in SCATTER/GATHER
-void insertScatterGatherSnippet(
-    ExecutionPlan& plan, ExecutionNode* at,
-    containers::SmallUnorderedMap<ExecutionNode*, ExecutionNode*> const&
-        subqueries);
-
-//// @brief find all subqueries in a plan and store a map from subqueries to
-/// nodes
-void findSubqueriesInPlan(
-    ExecutionPlan& plan,
-    containers::SmallUnorderedMap<ExecutionNode*, ExecutionNode*>& subqueries);
 
 void joinIndexNodesRule(Optimizer*, std::unique_ptr<ExecutionPlan>,
                         OptimizerRule const&);
